some bum ass flight code but whatever

+++ carm-electronics/flight-computer/utils.h
@@ -20,14 +20,14 @@
  * Returns: Nothing
  * Notes: The GPS module outputs the
  */
-double convertToDecimalDegrees(double coordinate, char direction)
+float convertToDecimalDegrees(float coordinate, char direction)
 {
     // extracting the degrees and minutes from the coordinate value
 
-    int degrees = (int)(coordinate / 100);    // first two digits represent degrees
-    double minutes = fmod(coordinate, 100.0); // remaining part are minutes
+    int degrees = (int)(coordinate / 100);   // first two digits represent degrees
+    float minutes = fmod(coordinate, 100.0); // remaining part are minutes
 
-    double decimal_degrees = (degrees + (minutes / 60.0));
+    float decimal_degrees = (degrees + (minutes / 60.0));
 
     // check the direction (N/S for latitude, E/W for longitude)
     if (direction == 'S' || direction == 'W')
